Implemented a sieve of Eratosthenes

+++ primelib/primelib.hpp
@@ -6,9 +6,9 @@ namespace primelib
 {
     bool check_prime_b(uint_fast64_t n_t);
 
-    void generatePrimeLUT(uint_fast64_t n, uint_fast64_t* &primeArray);
+    void generatePrimeLUT(uint_fast64_t n, uint_fast32_t* &primeArray);
 
-    uint_fast64_t retrievePrime(uint_fast64_t n, uint_fast64_t* primeArray);
+    uint_fast64_t retrievePrime(uint_fast64_t n, uint_fast32_t* primeArray);
 }
